SonarQube-driven cleanup

+++ Assets/Scripts/Base/Component/PowerUps/PowerUp.cs
@@ -58,7 +58,7 @@ namespace PowerUps
 
                 if (currentBubbleCount < lastBubbleCount)
                 {
-                    var fillRate = ((float)(max - current) - progress) / Mathf.Max(1, currentBubbleCount);
+                    var fillRate = ((max - current) - progress) / Mathf.Max(1, currentBubbleCount);
                     progress += (lastBubbleCount - currentBubbleCount) * fillRate;
                 }
